Implements a proper Google Analytics tag, and moves it to the header as trackign data should be submit before the <body> render process kicks off so we get in page analytics.
It also tracks slow load times, and other interesting bits if we include
Google webmaster toolkit - which we should add a low priority todo.

I'll be sending out access emails with this - if you need access to the
GA make sure you contact antonio, or the current team lead and they will
get you on the roster.